Add scaling method to improve fit (#13)
The method fit_baseline will estimate the empirical mean and covariance. Using fit_scaling will scale the covariance matrix in order to maximise the likelihood.

The scaling isn't very efficient right now because of repeated integration so compute the normalizing constant. In the future I plan to sample from a minimum volume ellipsoid enclosing the data which should significantly improve the efficiency and the results.

This is not yet working: https://github.com/FriesischScott/MinimumVolumeEllipsoids.jl